ipl-mlops/pyproject.toml
[tool.poetry]
name = "ipl-mlops"
version = "0.1.0"
description = "IPL Match Winner Prediction using Random Forest"
authors = ["Shankara Narayana N G <shankarnarayana92@gmail.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
scikit-learn = "^1.4.0"
joblib = "^1.3.0"
pyyaml = "^6.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
scipy = "^1.11.0"
jinja2 = "^3.1.0"
python-multipart = "^0.0.9"
pydantic = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
